requirements.txt
Flask==3.0.3
gunicorn==22.0.0
requests==2.32.3
wikipedia==1.4.0
googlesearch-python==1.2.5
pytz==2025.1